requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
streamlit
openai
pypdf